requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
requests-html==0.10.0
pandas==2.1.3
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
PyMySQL==1.1.0
